+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### New features
 
+* Added ability to ignore entire directories through .rubocop.yml
 * Relax semicolon rule for one line methods, classes and modules
 * New cop ClassMethods checks for uses for class/module names in definitions of class/module methods
 * New cop SingleLineMethods checks for methods implemented on a single line